+++ zuko-ui/src/components/PostV2/PostV2.jsx
@@ -1,4 +1,5 @@
-import styles from './HomePost.module.scss';
+import { NavLink } from 'react-router-dom';
+import styles from './PostV2.module.scss';
 
 const initialValue = {
 	id: 0,
@@ -9,9 +10,10 @@ const initialValue = {
 		'Bits of moving fluff white dwarf kindling the energy hidden in matter shores of the cosmic ocean Jean-François Champollion consciousness.',
 	coverImageUri: './placeholder.png',
 	category: 'KOLIKI JE MOJ DEO?',
+	readMoreUri: '#',
 };
 
-export default function HomePost({ postData = initialValue }) {
+export default function HomePostV2({ postData = initialValue }) {
 	return (
 		<div>
 			<div className={styles.post}>
@@ -35,6 +37,14 @@ export default function HomePost({ postData = initialValue }) {
 						</p>
 					)}
 				</div>
+				{postData.readMoreUri && (
+					<div className={styles.post__link}>
+						<NavLink to={postData.readMoreUri}>
+							Pročitaj više
+						</NavLink>
+						<img src="./arrow-right-orange.svg" alt='Arrow icon' />
+					</div>
+				)}
 			</div>
 		</div>
 	);
