+++ zuko-ui/src/pages/news.page.js
@@ -1,9 +1,9 @@
-import HomePostsContainer from '../components/HomePostsContainer/HomePostsContainer'
+import PostsContainer from '../components/PostsContainer/PostsContainer'
 import SearchPostInput from '../components/SearchPostInput/SearchPostInput'
 
 export default function NewsPage() {
     return <div>
 		<SearchPostInput />
-		<HomePostsContainer displayVersion="v2" />
+		<PostsContainer displayVersion="v2" />
     </div>
 }
